rename kadai1 -> kadai2
