--- kadai1/nagaa052/go.mod
@@ -1,1 +1,0 @@
-module github.com/gopherdojo/dojo5/kadai1/nagaa052
