+++ kadai2/nagaa052/Makefile
