+++ kadai2/nagaa052/go.mod
@@ -1,0 +1,1 @@
+module github.com/gopherdojo/dojo5/kadai2/nagaa052
