+++ kadai2/nagaa052/internal/gocon/format.go
