+++ kadai2/nagaa052/internal/gocon/format_test.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/gopherdojo/dojo5/kadai1/nagaa052/internal/gocon"
+	"github.com/gopherdojo/dojo5/kadai2/nagaa052/internal/gocon"
 )
 
 func TestGetExtentions(t *testing.T) {
