+++ kadai2/nagaa052/internal/gocon/gocon_test.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/gopherdojo/dojo5/kadai1/nagaa052/internal/gocon"
+	"github.com/gopherdojo/dojo5/kadai2/nagaa052/internal/gocon"
 )
 
 func TestRun(t *testing.T) {
