+++ kadai2/nagaa052/internal/gocon/testdata/TestRun/dirA/1.jpg
