+++ kadai2/nagaa052/pkg/convert/convert_test.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/gopherdojo/dojo5/kadai1/nagaa052/pkg/convert"
+	"github.com/gopherdojo/dojo5/kadai2/nagaa052/pkg/convert"
 )
 
 func TestToJpeg(t *testing.T) {
