+++ kadai2/nagaa052/pkg/search/search_test.go
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/gopherdojo/dojo5/kadai1/nagaa052/pkg/search"
+	"github.com/gopherdojo/dojo5/kadai2/nagaa052/pkg/search"
 )
 
 func TestWalk(t *testing.T) {
